+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 *Breaking*
 
-* `xan select -n` will not error anymore on empty inputs.
+* `xan select -n` will not error anymore on empty inputs and, generally, empty files should not trigger selection errors when using commands with `-n/--no-headers`.
 * `xan heatmap -C/--cram` becomes a flag accepting either `auto`, `always` or `never`.
 * Dropping `-C` short flag for `xan sort --cells` (it could be confused with `--columns` or `--check`).
 * Completely overhauled how datetimes work in moonblade.